Start event handlers in separate tasks
This has the potential for extremely unreadable logs and possibly race
conditions, but I'll deal with that if it pops up.

One possible solution is waiting for the task for a certain amount of
time before proceeding, which should let 99.99% of events get processed
in the right order without permanently blocking in corner cases.